Just some code cleanup.

+++ klgists/misc/prompts.py
@@ -1,7 +1,6 @@
 import shutil
 import os
 from enum import Enum
-from typing import Callable
 from colorama import Fore, Style
 
 
@@ -26,9 +25,19 @@ def prompt_and_delete(
 		trash_dir: str = os.path.join(os.environ['HOME'], '.Trash'),
 		allow_dirs: bool=True,
 		show_confirmation: bool=True,
-		test_only: bool=False,
+		dry: bool=False,
 		allow_ignore: bool=True
 ) -> Deletion:
+	"""
+
+	:param path:
+	:param trash_dir: The directory of a trashbin ('~/.Trash' by default)
+	:param allow_dirs: Allow deleting full directory trees; set to False for safety
+	:param show_confirmation: Print the action to stdout (ex: 'Trashed abc.txt to ~/.Trash')
+	:param dry: If True, will only return the Deletion object to be handled outside
+	:param allow_ignore: Allow entering an empty string to mean ignore
+	:return: A Deletion enum reflecting the chosen action
+	"""
 
 	assert allow_dirs or not os.path.isdir(path), 'Cannot delete directory {}; only files are allowed'.format(path)
 
@@ -38,12 +47,14 @@ def prompt_and_delete(
 
 		if command.lower() == Deletion.HARD.name.lower():
 			if show_confirmation: print(Style.BRIGHT + "Permanently deleted {}".format(path))
-			if os.path.isdir(path): shutil.rmtree(path)
-			else: os.remove(path)
+			if not dry:
+				if os.path.isdir(path): shutil.rmtree(path)
+				else: os.remove(path)
 			return Deletion.HARD
 
 		elif command.lower() == Deletion.TRASH.name.lower():
-			shutil.move(path, trash_dir)
+			if not dry:
+				shutil.move(path, trash_dir)
 			if show_confirmation: print(Style.BRIGHT + "Trashed {} to {}".format(path, trash_dir))
 			return Deletion.TRASH
